services/backend/requirements.txt
fastapi
uvicorn[standard]
httpx
sqlmodel
sqlalchemy
python-dotenv
google-generativeai
pydantic
